requirements.txt
flask==3.0.0
gunicorn==21.2.0
line-bot-sdk==3.5.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gspread==5.12.0
python-dotenv==1.0.0
supabase==2.10.0
